requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
chromadb
gitpython
python-dotenv
langchain-ollama
python-frontmatter
requests
